# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: VouckBack
  description: |-
    Microserviço para geração de voucher como descontos em valor ou percentual paras clientes em seu ponto de venda.

    _Se você que saber mais sobre a TIBBI, então  click [aqui](http://tibbi.com.br)._

  contact:
    email: mario7lagoas@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8080/voucher-back/v1
...
paths:
...
  /usuario:
    get:
      tags:
        - Usuario
      summary: Buscando Lista de Usuario
      description: Buscando Lista de Usuario na Plataforma
      operationId: buscandoListaUsuario
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioApiResponse'
    post:
      tags:
        - Usuario
      summary: Criando Um Usuario
      description: Criando Um Usuario na Plataforma
      operationId: criandoUsuario
      requestBody:
        description: Criando novo Usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioApiRequest'
        required: true
      responses:
        '201':
          description: created successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioApiResponse'
        '400':
          description: bad request operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 400 BAD_REQUEST
                  erros:
                    type: array
                    items:
                      properties:
                        mensagem:
                          type: string
                          example: "não deve ser nulo"
                        campo:
                          type: string
                          example: "email"
        '422':
          description: unprocessable operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 422 UNPROCESSABLE_ENTITY
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: "Usuario já Cadastrado."
                        mensagem:
                          type: string
                          example: "E-mail já cadastrado."
  /usuario/{guid}:
    get:
      tags:
        - Usuario
      summary: Buscando Usuario Pelo GUID
      description: Buscando Usuario Pelo GUID na Plataforma
      operationId: buscandoUsuarioPeloGUID
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Usuario não encontrado."
    put:
      tags:
        - Usuario
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          example: aaaa-bbbb-cccc-dddd-eeeee
          required: true
          schema:
            type: string
      summary: Alterando um Usuario
      description: Alterando um LUsuario na Plataforma
      operationId: alterandoUsuario
      requestBody:
        description: Alterando Usuario Existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioUpdateApiRequest'
      responses:
        '202':
          description: accepted successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Usuario não encontrado."
    delete:
      tags:
        - Usuario
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      summary: Apagando um Usuario
      description: Apagando um Usuario na Plataforma
      operationId: apagandoUsuario
      responses:
        '204':
          description: No Content successful operation
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Usuario não encontrado."
    patch:
      tags:
        - Usuario
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      summary: Alterando Status de um Usuario
      description: Alterando Status de um Usuario na Plataforma
      operationId: alterandoStatusUsuario
      requestBody:
        description: Alterando Usuario Existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusApiRequest'
        required: true
      responses:
        '202':
          description: Accepted successful operation
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Usuario não encontrado."
  /usuario/paginada:
    get:
      tags:
        - Usuario
      summary: Buscando Lista de Usuario Pagina
      description: Buscando Lista de Usuario pagina na Plataforma
      operationId: buscandoListaPaginadaUsuario
      parameters:
        - name: page
          description: "Numero da pagina atual"
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          description: "Quantidade de registro por pagina"
          in: query
          schema:
            type: integer
            default: 10
        - name: nome
          description: "Nome do Usuario"
          in: query
          schema:
            type: string
            default: ""
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginacaoApiResponse'
                  - type: object
                    properties:
                      usuarios:
                        type: array
                        items:
                          $ref: '#/components/schemas/UsuarioApiResponse'
  /usuario/print:
    post:
      tags:
        - Usuario
      summary: Relatorio de Usuario
      description: Relatorio de Usuario na Plataforma
      operationId: relatorioUsuario
      requestBody:
        description: Imprimindo um relatorio de Usuario
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/UsuarioApiResponse'
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: "data:application/pdf;base64,JVBERi0xLjUKJeLjz9MKMyAwIG9iago8PC9Db2xvclNwYWNlL0RldmljZVJHQi9TdWJ0eXBlL0ltYWdlL0hlaWdodCAzNTQvRmlsdGVyL0RDVERlY29kZS9UeXBlL1hPYmplY3QvV2lkdGggMjQ4MC9CaXRzUGVyQ29tcG9uZW50IDgvTGVuZ3RoIDM2ODg5Pj5zdHJlYW0K/9j/4AAQSkZJRgABAgAAZABkAAD/7AARRHVja3kAAQAEAAAACgAA/+4ADkFkb2JlAGTAAAAAAf/bAIQAFBAQGRIZJxcXJzImHyYyLiYmJiYuPjU1NTU1PkRBQUFBQUFERERERERERERERERERERERERERERERERERERERAEVGRkgHCAmGBgmNiYgJjZENisrNkREREI1QkRERERERERERERERERERERERERERERERERERERERERERERERERE/8AAEQgBYgmwAwEiAAIRAQMRAf/..."
components:
  schemas:
    PaginacaoApiResponse:
      properties:
        number:
          type: integer
          example: 0
          description: "Pagina atual"
        totalPages:
          type: integer
          example: 1
          description: "Total de paginas"
        totalElements:
          type: integer
          example: 1
          description: "Total de Items"
    GuidApiRequest:
      required:
        - guid
      type: object
      properties:
        guid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador guid"
    UpdateStatusApiRequest:
      required:
        - status
      type: object
      properties:
        status:
          type: boolean
          description: "Status [true/false]"
          example: true
...
    LojaApiResponse:
      properties:
        guid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador guid"
        nome:
          type: string
          example: "Loja 001"
          description: "Nome da Loja"
        cnpj:
          type: string
          example: "4249270100013"
          description: "CNPJ da Loja"
        identificacao:
          type: string
          example: "0001"
          description: "Idenficação da Loja"
        status:
          type: boolean
          example: true
          description: "Status da Loja"
        dataCadastro:
          type: string
          example: "2024-05-07 11:48:11"
          description: "Data do cadastro"
...
    PerfilResumidoApiResponse:
      type: object
      properties:
        guid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador guid"
        nome:
          type: string
          example: "Moderator"
          description: "Nome do Perfil"
...
    UsuarioApiResponse:
      properties:
        guid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador guid"
        userName:
          type: string
          description: "Nome do Usuario"
          example: "Mario Dias"
        email:
          type: string
          description: "E-mail do Usuario"
          example: "mario7lagoas@gmail.com"
        status:
          type: boolean
          description: "Status do Usuario"
          example: true
        dataCadastro:
          type: string
          description: "Data do Cadastro do Usuario"
          example: "2024-01-01 12:00:00"
        dataAtualizacao:
          type: string
          description: "Data do Atualização do Usuario"
          example: "2024-01-01 13:20:00"
        perfis:
          type: array
          description: "Permissões do Usuario"
          items:
            $ref: '#/components/schemas/PerfilResumidoApiResponse'
        lojas:
          type: array
          description: "Lojas associadas ao Usuario"
          items:
            $ref: '#/components/schemas/LojaApiResponse'
    UsuarioPerfilApiRequest:
      type: object
      properties:
        nome:
          type: string
          example: "Moderator"
          description: "Perfil do Usuario"
    UsuarioApiRequest:
      required:
        - userName
        - email
        - status
        - password
        - perfis
      type: object
      properties:
        userName:
          type: string
          description: "Nome do Usuario"
          example: "Mario Dias"
        email:
          type: string
          description: "E-mail do Usuario"
          example: "mario7lagoas@gmail.com"
        password:
          type: string
          description: "Senha do Usuario"
          example: "123456"
        status:
          type: boolean
          description: "Status do Usuario"
          example: true
        perfis:
          type: array
          description: "Perfil do Usuario"
          items:
            $ref: '#/components/schemas/UsuarioPerfilApiRequest'
        lojas:
          type: array
          description: "Lojas associadas ao Usuario"
          items:
            $ref: '#/components/schemas/GuidApiRequest'
    UsuarioUpdateApiRequest:
      type: object
      properties:
        userName:
          type: string
          description: "Nome do Usuario"
          example: "Mario Dias"
        email:
          type: string
          description: "E-mail do Usuario"
          example: "mario7lagoas@gmail.com"
        password:
          type: string
          description: "Senha do Usuario"
          example: "123456"
        status:
          type: boolean
          description: "Status do Usuario"
          example: true
        perfis:
          type: array
          description: "Perfil do Usuario"
          items:
            $ref: '#/components/schemas/UsuarioPerfilApiRequest'
        lojas:
          type: array
          description: "Lojas associadas ao Usuario"
          items:
            $ref: '#/components/schemas/GuidApiRequest'